Honor fractional delays and pace every reaction, thread, and emoji request
The delay selectors used parseInt, so a 0.5s search delay ran loops with
no delay and 2.9s ran as 2s while the Settings caption showed the fraction.
Parse as floats.

Three loops fired requests back to back: purge reactions mode between
emojis and messages, export thread walks between threads, and batch
reaction removal between emojis on one message. Each now sleeps on the
configured delay, with the invariant under test that no two requests run
without a sleep between them.

+++ src/features/app/appSlice.ts
@@ -234,11 +234,14 @@ export const selectSidebarView = (state: RootState) => state.app.sidebarView;
 export const selectSetting = (key: DiscrubSetting) => (state: RootState) =>
   state.app.settings?.[key] ?? defaultSettings[key];
 
+// The delay sliders store one decimal ("0.5", "2.9"). These MUST parse as
+// floats: parseInt silently truncated "0.5" to 0 and "2.9" to 2, so every
+// operation loop slept less than the Settings tab showed (2.1.3 pacing audit).
 export const selectSearchDelay = (state: RootState) =>
-  parseInt(state.app.settings?.[DiscrubSetting.SEARCH_DELAY] ?? defaultSettings[DiscrubSetting.SEARCH_DELAY]);
+  parseFloat(state.app.settings?.[DiscrubSetting.SEARCH_DELAY] ?? defaultSettings[DiscrubSetting.SEARCH_DELAY]);
 
 export const selectDeleteDelay = (state: RootState) =>
-  parseInt(state.app.settings?.[DiscrubSetting.DELETE_DELAY] ?? defaultSettings[DiscrubSetting.DELETE_DELAY]);
+  parseFloat(state.app.settings?.[DiscrubSetting.DELETE_DELAY] ?? defaultSettings[DiscrubSetting.DELETE_DELAY]);
 
 export const selectDelayModifier = (state: RootState) =>
   parseFloat(state.app.settings?.[DiscrubSetting.DELAY_MODIFIER] ?? defaultSettings[DiscrubSetting.DELAY_MODIFIER]);
